+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-3.x/.template.config/vs-2017.3.host.json
@@ -1,0 +1,47 @@
+{
+  "name": {
+    "text": "Cosmos DB Trigger",
+    "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}",
+    "id": "2010"
+  },
+  "description": {
+    "text": "A C# function that will be run whenever documents change in a document collection.",
+    "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}",
+    "id": "2035"
+  },
+  "order": 700,
+  "icon": "vs-2017.3/cosmosdb.png",
+  "learnMoreLink": "https://go.microsoft.com/fwlink/?linkid=848145",
+  "uiFilters": [
+    "functions"
+  ],
+  "symbolInfo": [
+      {
+        "id": "ConnectionStringSetting",
+        "name": {
+          "id": "2011",
+          "text": "Connection string setting name",
+          "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}"
+        },
+        "isVisible": true
+      },
+      {
+        "id": "DatabaseName",
+        "name": {
+          "id": "2012",
+          "text": "Database name",
+          "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}"
+        },
+        "isVisible": true
+      },
+      {
+        "id": "CollectionName",
+        "name": {
+          "id": "2013",
+          "text": "Collection name",
+          "package": "{E9BCFE2B-304A-4D7F-836F-F3A4FB1A1451}"
+        },
+        "isVisible": true
+      }
+  ]
+}
