+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-3.x/.template.config/vs-2017.3/cosmosdb-mac-32~dark~sel@2x.png
