+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-3.x/.template.config/vs-2017.3/cosmosdb-mac-32~sel.png
