+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-3.x/function.json
@@ -1,0 +1,14 @@
+{
+  "bindings": [
+    {
+      "type": "cosmosDBTrigger",
+      "name": "input",
+      "direction": "in",
+      "connectionStringSetting": "",
+      "databaseName": "",
+      "collectionName": "",
+      "leaseCollectionName": "leases",
+      "createLeaseCollectionIfNotExists" : true
+    }
+  ]
+}
