+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-3.x/metadata.json
@@ -2,7 +2,7 @@
   "defaultFunctionName": "CosmosTrigger",
   "description": "$CosmosDBTrigger_description",
   "name": "Azure Cosmos DB trigger",
-  "language": "Custom",
+  "language": "C#",
   "category": [
     "$temp_category_core",
     "$temp_category_dataProcessing"
@@ -10,10 +10,10 @@
   "categoryStyle": "cosmosDB",
   "enabledInTryMode": false,
   "userPrompt": [
-    "connection",
+    "connectionStringSetting",
     "databaseName",
-    "containerName",
-    "leaseContainerName",
-    "createLeaseContainerIfNotExists"
+    "collectionName",
+    "leaseCollectionName",
+    "createLeaseCollectionIfNotExists"
   ]
 }
