+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-4.x/CosmosDBTriggerCSharp.cs
@@ -1,0 +1,33 @@
+using System;
+using System.Collections.Generic;
+using Microsoft.Azure.WebJobs;
+using Microsoft.Azure.WebJobs.Host;
+using Microsoft.Extensions.Logging;
+
+namespace Company.Function
+{
+    public static class CosmosDBTriggerCSharp
+    {
+        [FunctionName("CosmosDBTriggerCSharp")]
+        public static void Run([CosmosDBTrigger(
+            databaseName: "DatabaseName",
+            containerName: "ContainerName",
+            Connection = "CosmosDBConnectionSetting",
+            LeaseContainerName = "leases")]IReadOnlyList<ToDoItem> input,
+            ILogger log)
+        {
+            if (input != null && input.Count > 0)
+            {
+                log.LogInformation("Documents modified " + input.Count);
+                log.LogInformation("First document Id " + input[0].id);
+            }
+        }
+    }
+
+    // Customize the model with your own desired properties
+    public class ToDoItem
+    {
+        public string id { get; set; }
+        public string Description { get; set; }
+    }
+}
