+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-4.x/function.json
@@ -2,13 +2,14 @@
   "bindings": [
     {
       "type": "cosmosDBTrigger",
-      "name": "documents",
+      "name": "input",
       "direction": "in",
       "leaseContainerName": "leases",
       "connection": "",
       "databaseName": "",
       "containerName": "",
-      "createLeaseContainerIfNotExists" : true
+      "createLeaseContainerIfNotExists": true,
+      "startFromBeginning": false
     }
   ]
 }
