+++ Functions.Templates/Templates/CosmosDBTrigger-CSharp-4.x/metadata.json
@@ -2,7 +2,7 @@
   "defaultFunctionName": "CosmosTrigger",
   "description": "$CosmosDBTrigger_description",
   "name": "Azure Cosmos DB trigger",
-  "language": "Custom",
+  "language": "C#",
   "category": [
     "$temp_category_core",
     "$temp_category_dataProcessing"
