--- Functions.Templates/Templates/CosmosDBTrigger-CSharp-5.x/CosmosDBTriggerCSharp.cs
@@ -1,27 +1,0 @@
-using System;
-using System.Collections.Generic;
-using Microsoft.Azure.Documents;
-using Microsoft.Azure.WebJobs;
-using Microsoft.Azure.WebJobs.Host;
-using Microsoft.Extensions.Logging;
-
-namespace Company.Function
-{
-    public static class CosmosDBTriggerCSharp
-    {
-        [FunctionName("CosmosDBTriggerCSharp")]
-        public static void Run([CosmosDBTrigger(
-            databaseName: "DatabaseValue",
-            collectionName: "CollectionValue",
-            ConnectionStringSetting = "ConnectionValue",
-            LeaseCollectionName = "leases")]IReadOnlyList<Document> input,
-            ILogger log)
-        {
-            if (input != null && input.Count > 0)
-            {
-                log.LogInformation("Documents modified " + input.Count);
-                log.LogInformation("First document Id " + input[0].Id);
-            }
-        }
-    }
-}
