--- Functions.Templates/Templates/CosmosDBTrigger-CSharp-5.x/function.json
@@ -1,15 +1,0 @@
-{
-  "bindings": [
-    {
-      "type": "cosmosDBTrigger",
-      "name": "input",
-      "direction": "in",
-      "leaseContainerName": "leases",
-      "connection": "",
-      "databaseName": "",
-      "containerName": "",
-      "createLeaseContainerIfNotExists": true,
-      "startFromBeginning": false
-    }
-  ]
-}
