--- Functions.Templates/Templates/CosmosDBTrigger-CSharp-5.x/run.csx
@@ -1,13 +1,0 @@
-#r "Microsoft.Azure.DocumentDB.Core"
-using System;
-using System.Collections.Generic;
-using Microsoft.Azure.Documents;
-
-public static void Run(IReadOnlyList<Document> input, ILogger log)
-{
-    if (input != null && input.Count > 0)
-    {
-        log.LogInformation("Documents modified " + input.Count);
-        log.LogInformation("First document Id " + input[0].Id);
-    }
-}
