--- Functions.Templates/Templates/CosmosDBTrigger-CSharp/.template.config/template.json
@@ -1,81 +1,0 @@
-{
-  "author": "Microsoft",
-  "classifications": [
-    "Azure Function",
-    "Trigger",
-    "Cosmos DB"
-  ],
-  "name": "CosmosDBTrigger",
-  "identity": "Azure.Function.CSharp.CosmosDBTrigger.2.x",
-  "groupIdentity": "Azure.Function.CosmosDBTrigger",
-  "shortName": "cosmos",
-  "tags": {
-    "language": "C#",
-    "type": "item"
-  },
-  "sourceName": "CosmosDBTriggerCSharp",
-  "symbols": {
-    "namespace": {
-      "description": "namespace for the generated code",
-      "replaces": "Company.Function",
-      "type": "parameter"
-    },
-    "ConnectionStringSetting": {
-      "description": "The name of the App Setting containing the connection string to the service that contains the collection to be monitored.",
-      "type": "parameter",
-      "defaultValue": "",
-      "replaces": "ConnectionValue"
-    },
-    "DatabaseName": {
-      "description": "This is the name of the database that includes the collection to be monitored.",
-      "type": "parameter",
-      "defaultValue": "databaseName",
-      "replaces": "DatabaseValue"
-    },
-    "CollectionName": {
-      "description": "This is the name of the collection to be monitored.",
-      "type": "parameter",
-      "defaultValue": "collectionName",
-      "replaces": "CollectionValue"
-    },
-    "HostIdentifier": {
-      "type": "bind",
-      "binding": "HostIdentifier"
-    }
-  },
-  "primaryOutputs": [
-    {
-      "path": "CosmosDBTriggerCSharp.cs"
-    }
-  ],
-  "defaultName": "CosmosDBTriggerCSharp",
-  "postActions": [
-    {
-      "Description": "Adding Reference to Microsoft.Azure.WebJobs.Extensions.CosmosDB Nuget package",
-      "ActionId": "B17581D1-C5C9-4489-8F0A-004BE667B814",
-      "ContinueOnError": "true",
-      "ManualInstructions": [],
-      "args": {
-        "referenceType": "package",
-        "reference": "Microsoft.Azure.WebJobs.Extensions.CosmosDB", "version": "3.0.10",
-        "projectFileExtensions": ".csproj"
-      }
-    },
-    {
-      "description": "Restore NuGet packages required by this project.",
-      "manualInstructions": [],
-      "actionId": "210D431B-A78B-4D2F-B762-4ED3E3EA9025",
-      "continueOnError": true
-    },
-    {
-      "condition": "(HostIdentifier != \"dotnetcli\" && HostIdentifier != \"dotnetcli-preview\")",
-      "description": "Opens the function class file in the editor",
-      "manualInstructions": [],
-      "actionId": "84C0DA21-51C8-4541-9940-6CA19AF04EE6",
-      "args": {
-        "files": "0"
-      },
-      "continueOnError": true
-    }
-  ]
-}
