--- Functions.Templates/Templates/CosmosDBTrigger-CSharp/function.json
@@ -1,14 +1,0 @@
-{
-  "bindings": [
-    {
-      "type": "cosmosDBTrigger",
-      "name": "input",
-      "direction": "in",
-      "connectionStringSetting": "",
-      "databaseName": "",
-      "collectionName": "",
-      "leaseCollectionName": "leases",
-      "createLeaseCollectionIfNotExists" : true
-    }
-  ]
-}
