--- Functions.Templates/Templates/CosmosDBTrigger-CSharp/metadata.json
@@ -1,19 +1,0 @@
-{
-  "defaultFunctionName": "CosmosTrigger",
-  "description": "$CosmosDBTrigger_description",
-  "name": "Azure Cosmos DB trigger",
-  "language": "C#",
-  "category": [
-    "$temp_category_core",
-    "$temp_category_dataProcessing"
-  ],
-  "categoryStyle": "cosmosDB",
-  "enabledInTryMode": false,
-  "userPrompt": [
-    "connectionStringSetting",
-    "databaseName",
-    "collectionName",
-    "leaseCollectionName",
-    "createLeaseCollectionIfNotExists"
-  ]
-}
