+++ Functions.Templates/Templates/CosmosDBTrigger-JavaScript-4.x/index.js
@@ -1,0 +1,5 @@
+module.exports = async function (context, documents) {
+    if (!!documents && documents.length > 0) {
+        context.log('Document Id: ', documents[0].id);
+    }
+}
