+++ Functions.Templates/Templates/CosmosDBTrigger-JavaScript-4.x/metadata.json
@@ -2,7 +2,7 @@
   "defaultFunctionName": "CosmosTrigger",
   "description": "$CosmosDBTrigger_description",
   "name": "Azure Cosmos DB trigger",
-  "language": "Custom",
+  "language": "JavaScript",
   "category": [
     "$temp_category_core",
     "$temp_category_dataProcessing"
