+++ Functions.Templates/Templates/CosmosDBTrigger-JavaScript-4.x/sample.dat
@@ -1,0 +1,5 @@
+[
+	{
+		"id": "sample"
+	}
+]
