--- Functions.Templates/Templates/CosmosDBTrigger-JavaScript-5.x/function.json
@@ -1,15 +1,0 @@
-{
-  "bindings": [
-    {
-      "type": "cosmosDBTrigger",
-      "name": "documents",
-      "direction": "in",
-      "leaseContainerName": "leases",
-      "connection": "",
-      "databaseName": "",
-      "containerName": "",
-      "createLeaseContainerIfNotExists": true,
-      "startFromBeginning": false
-    }
-  ]
-}
