--- Functions.Templates/Templates/CosmosDBTrigger-JavaScript-5.x/index.js
@@ -1,5 +1,0 @@
-module.exports = async function (context, documents) {
-    if (!!documents && documents.length > 0) {
-        context.log('Document Id: ', documents[0].id);
-    }
-}
