--- Functions.Templates/Templates/CosmosDBTrigger-JavaScript-5.x/sample.dat
@@ -1,5 +1,0 @@
-[
-	{
-		"id": "sample"
-	}
-]
