+++ Functions.Templates/Templates/CosmosDBTrigger-PowerShell-4.x/function.json
@@ -8,7 +8,8 @@
       "connection": "",
       "databaseName": "",
       "containerName": "",
-      "createLeaseContainerIfNotExists" : true
+      "createLeaseContainerIfNotExists": true,
+      "startFromBeginning": false
     }
   ]
 }
