+++ Functions.Templates/Templates/CosmosDBTrigger-PowerShell-4.x/metadata.json
@@ -2,7 +2,7 @@
   "defaultFunctionName": "CosmosTrigger",
   "description": "$CosmosDBTrigger_description",
   "name": "Azure Cosmos DB trigger",
-  "language": "Custom",
+  "language": "PowerShell",
   "category": [
     "$temp_category_core",
     "$temp_category_dataProcessing"
