+++ Functions.Templates/Templates/CosmosDBTrigger-PowerShell-4.x/run.ps1
