+++ Functions.Templates/Templates/CosmosDBTrigger-PowerShell-4.x/sample.dat
@@ -1,0 +1,5 @@
+[
+	{
+		"id": "sample"
+	}
+]
