--- Functions.Templates/Templates/CosmosDBTrigger-PowerShell-5.x/metadata.json
@@ -1,19 +1,0 @@
-{
-  "defaultFunctionName": "CosmosTrigger",
-  "description": "$CosmosDBTrigger_description",
-  "name": "Azure Cosmos DB trigger",
-  "language": "PowerShell",
-  "category": [
-    "$temp_category_core",
-    "$temp_category_dataProcessing"
-  ],
-  "categoryStyle": "cosmosDB",
-  "enabledInTryMode": false,
-  "userPrompt": [
-    "connection",
-    "databaseName",
-    "containerName",
-    "leaseContainerName",
-    "createLeaseContainerIfNotExists"
-  ]
-}
