--- Functions.Templates/Templates/CosmosDBTrigger-PowerShell-5.x/sample.dat
@@ -1,5 +1,0 @@
-[
-	{
-		"id": "sample"
-	}
-]
