+++ Functions.Templates/Templates/CosmosDBTrigger-Python-4.x/sample.dat
@@ -1,0 +1,5 @@
+[
+	{
+		"id": "sample"
+	}
+]
