--- Functions.Templates/Templates/CosmosDBTrigger-Python-5.x/__init__.py
@@ -1,8 +1,0 @@
-import logging
-
-from azure.functions import DocumentList
-
-
-def main(documents: DocumentList) -> str:
-    if documents:
-        logging.info('Document id: %s', documents[0]['id'])
