--- Functions.Templates/Templates/CosmosDBTrigger-Python-5.x/function.json
@@ -1,16 +1,0 @@
-{
-  "scriptFile": "__init__.py",
-  "bindings": [
-    {
-      "type": "cosmosDBTrigger",
-      "name": "documents",
-      "direction": "in",
-      "leaseContainerName": "leases",
-      "connection": "",
-      "databaseName": "",
-      "containerName": "",
-      "createLeaseContainerIfNotExists": true,
-      "startFromBeginning": false
-    }
-  ]
-}
