--- Functions.Templates/Templates/CosmosDBTrigger-Python-5.x/metadata.json
@@ -1,22 +1,0 @@
-{
-  "defaultFunctionName": "CosmosTrigger",
-  "description": "$CosmosDBTrigger_description",
-  "name": "Azure Cosmos DB trigger",
-  "language": "Python",
-  "category": [
-    "$temp_category_core",
-    "$temp_category_dataProcessing"
-  ],
-  "categoryStyle": "cosmosDB",
-  "enabledInTryMode": false,
-  "userPrompt": [
-    "connection",
-    "databaseName",
-    "containerName",
-    "leaseContainerName",
-    "createLeaseContainerIfNotExists"
-  ],
-  "filters": [
-    "Python3"
-  ]
-}
